// src/enums/project.enums.ts
export enum EProject {
    MAXEBCOUNT = 'maxEbCOunt',
    MINEBCOUNT = 'minEbCOunt',
    MAXFOLLOWERS = 'maxFollowers',
    MINFOLLOWERS = 'minFollowers',
    MAXTWEETS = 'maxTweets',
    MINTWEETS = 'minTweets',
    MAX24HCHANGE = 'max24hChange',
    MIN24HCHANGE = 'min24hChange',
    MAX7DCHANGE = 'max7dChange',
    MIN7DCHANGE = 'min7dChange',
}